my-telegram-bot/requirements.txt
telebot
google-generativeai
Pillow
requests